+++ code/header/model/sg_models/exciter_model/PSASPE14.h
@@ -82,7 +82,7 @@ class PSASPE14: public EXCITER_MODEL
 
         FIRST_ORDER_BLOCK sensor, regulator, ifd_feedback, rectifier;
         PI_BLOCK regulator_pi, current_pi;
-        double Efdmax, Efdmin;
+        float Efdmax, Efdmin;
 };
 
 #endif // EXCITER_MODEL_H
