+++ code/source/device/load.cpp
@@ -9,8 +9,8 @@
 
 using namespace std;
 
-double LOAD::voltage_threshold_of_constant_power_load_in_pu = 0.7;
-double LOAD::voltage_threshold_of_constant_current_load_in_pu = 0.5;
+float LOAD::voltage_threshold_of_constant_power_load_in_pu = 0.7;
+float LOAD::voltage_threshold_of_constant_current_load_in_pu = 0.5;
 
 LOAD::LOAD(STEPS& toolkit) : DEVICE(toolkit)
 {
